pyproject.toml
[project]
name = "job-recommended-with-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.2.0",
    "langchain-groq>=0.3.8",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
